Added missing library in pom.xml. Set target java version for IntelliJ users.
Apparently junit-jupiter-params was missing in pom configurations, althrough this may not cause problems to Eclipse users, who import this project for the first time, it will certainly be one for IntelliJ users, who depends on maven configuration.

Also I have set target java version to 8 due to IntelliJ's default version be set to 5 by default if it failed in find it in pom.xml.